requirements/local.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
ConfigArgParse==1.5.3
decorator==5.1.1
Django==4.1.6
django-extensions==3.2.1
django-ipware==4.0.2
django-shell-plus==1.1.7
django-structlog[celery]==4.1.1
django-tastypie==0.14.5
executing==1.2.0
Faker==16.8.1
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.2.0
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.2
humanize==4.6.0
idna==3.4
ipdb==0.13.11
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kombu==5.2.4
locust==2.14.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
msgpack==1.0.4
mysqlclient==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-mimeparse==1.6.0
pytz==2022.7.1
pyzmq==25.0.0
redis==4.5.1
requests==2.28.2
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
structlog==22.3.0
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
zope.event==4.6
zope.interface==5.5.2
